# model: make ID fields non-pointer string types