--- a/model/modeldecoder/transaction.go
+++ b/model/modeldecoder/transaction.go
@@ -104,12 +104,12 @@ func decodeRUMV3Spans(raw map[string]interface{}, input Input, tr *model.Transac
 		if err != nil {
 			return spans, err
 		}
-		span.TransactionID = &tr.ID
-		span.TraceID = &tr.TraceID
+		span.TransactionID = tr.ID
+		span.TraceID = tr.TraceID
 		if span.ParentIdx == nil {
-			span.ParentID = &tr.ID
+			span.ParentID = tr.ID
 		} else if *span.ParentIdx < idx {
-			span.ParentID = &spans[*span.ParentIdx].ID
+			span.ParentID = spans[*span.ParentIdx].ID
 		}
 		spans[idx] = span
 	}
@@ -158,12 +158,12 @@ func decodeTransaction(input Input, schema *jsonschema.Schema) (*model.Transacti
 		SpanCount: model.SpanCount{
 			Dropped: decoder.IntPtr(raw, fieldName("dropped"), fieldName("span_count")),
 			Started: decoder.IntPtr(raw, fieldName("started"), fieldName("span_count"))},
-		ParentID: decoder.StringPtr(raw, fieldName("parent_id")),
-		TraceID:  decoder.String(raw, fieldName("trace_id")),
+		TraceID: decoder.String(raw, fieldName("trace_id")),
 	}
 	if decoder.Err != nil {
 		return nil, decoder.Err
 	}
+	decodeString(raw, fieldName("parent_id"), &e.ParentID)
 	if e.Timestamp.IsZero() {
 		e.Timestamp = input.RequestTime
 	}