--- a/model/modeldecoder/transaction_test.go
+++ b/model/modeldecoder/transaction_test.go
@@ -331,7 +331,7 @@ func TestTransactionEventDecode(t *testing.T) {
 				Type:      trType,
 				Name:      &name,
 				Result:    &result,
-				ParentID:  &parentID,
+				ParentID:  parentID,
 				TraceID:   traceID,
 				Duration:  duration,
 				Timestamp: timestampParsed,