--- a/model/transaction.go
+++ b/model/transaction.go
@@ -44,7 +44,7 @@ type Transaction struct {
 	Metadata Metadata
 
 	ID       string
-	ParentID *string
+	ParentID string
 	TraceID  string
 
 	Timestamp time.Time
@@ -121,7 +121,7 @@ func (e *Transaction) Transform(ctx context.Context, tctx *transform.Context) []
 
 	// then merge event specific information
 	utility.AddId(fields, "parent", e.ParentID)
-	utility.AddId(fields, "trace", &e.TraceID)
+	utility.AddId(fields, "trace", e.TraceID)
 	utility.Set(fields, "timestamp", utility.TimeAsMicros(e.Timestamp))
 	// merges with metadata labels, overrides conflicting keys
 	utility.DeepUpdate(fields, "labels", e.Labels.Fields())