… model/modeldecoder/error_test.go
… model/modeldecoder/span_test.go