## Motivation/summary

Make more ID fields non-pointer types.

I missed a subtle issue with https://github.com/elastic/apm-server/pull/3849, which changed some code to store a pointer to an empty parent ID in some cases. This didn't cause any change in the fields output, because we check `!= nil` *and* `!= ""` when producing beat events. However, when aggregating transaction metrics, we are only checking `!= nil`.

To simplify the logic, and avoid these kinds of errors in the future, just store a string value. As a side benefit, this reduces heap allocations.

```
$ benchstat /tmp/old.txt /tmp/new.txt
name                                          old time/op    new time/op    delta                                                                                                                          
BackendProcessor/heavy.ndjson/NoRateLimit-12    38.4ms ± 4%    33.2ms ± 7%  -13.51%  (p=0.008 n=5+5)                                                                                                       
                                                                                                                                                                                                           
name                                          old speed      new speed      delta                                                                                                                          
BackendProcessor/heavy.ndjson/NoRateLimit-12  10.4MB/s ± 4%  12.0MB/s ± 7%  +15.69%  (p=0.008 n=5+5)                                                                                                       
                                                                                                                                                                                                           
name                                          old alloc/op   new alloc/op   delta                                                                                                                          
BackendProcessor/heavy.ndjson/NoRateLimit-12    6.34MB ± 0%    6.33MB ± 0%     ~     (p=0.151 n=5+5)
                                                                                                     
name                                          old allocs/op  new allocs/op  delta
BackendProcessor/heavy.ndjson/NoRateLimit-12      113k ± 0%      112k ± 0%   -1.27%  (p=0.008 n=5+5)
```

Included is a tangentially related change to stop recording the parent span index, used for RUMv3, on the model types. We now just use the index during decoding and then discard it.

## Checklist

- [x] I have signed the [Contributor License Agreement](https://www.elastic.co/contributor-agreement/).
~- [ ] I have updated [CHANGELOG.asciidoc](https://github.com/elastic/apm-server/blob/master/CHANGELOG.asciidoc)~

I have considered changes for:
~- [ ] documentation~
~- [ ] logging (add log lines, choose appropriate log selector, etc.)~
~- [ ] metrics and monitoring (create issue for Kibana team to add metrics to visualizations, e.g. [Kibana#44001](https://github.com/elastic/kibana/issues/44001))~
- [x] automated tests (add tests for the code changes, all [**unit** tests](https://github.com/elastic/apm-server/blob/master/TESTING.md) pass locally)
~- [ ] telemetry~
~- [ ] Elasticsearch Service (https://cloud.elastic.co)~
~- [ ] Elastic Cloud Enterprise (https://www.elastic.co/products/ece)~
~- [ ] Elastic Cloud on Kubernetes (https://www.elastic.co/elastic-cloud-kubernetes)~

## How to test these changes

`make test`

## Related issues

None.